backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-JWT-Extended
python-dotenv
pymysql
Werkzeug
